sbt-test-order-dependency-finder: notice failing pairs in O(1) runs
If a pair of test cases are sufficient to provoke the failure, this
will be noticed within the first two runs, then shrinking can begin.

This is accomplished by trying every pair `{s, t}` of test cases in
both orderings in the first two runs.

Larger k-tuples are tried after expected, not guaranteed, k! runs.